cleanup OCRPrivateKey
