+++ core/store/models/ocrkey/off_chain_public_key.go
@@ -1,0 +1,5 @@
+package ocrkey
+
+import "crypto/ed25519"
+
+type OffChainPublicKey ed25519.PublicKey
