+++ core/store/models/ocrkey/on_chain_public_key.go
@@ -1,0 +1,13 @@
+package ocrkey
+
+import (
+	"crypto/ecdsa"
+
+	"github.com/ethereum/go-ethereum/crypto"
+)
+
+type OnChainPublicKey ecdsa.PublicKey
+
+func (k OnChainPublicKey) Address() OnChainSigningAddress {
+	return OnChainSigningAddress(crypto.PubkeyToAddress(ecdsa.PublicKey(k)))
+}
