+++ core/store/models/ocrkey/private_key.go
@@ -4,11 +4,9 @@ import (
 	"crypto/ecdsa"
 	"crypto/ed25519"
 	"crypto/sha256"
-	"database/sql/driver"
 	"encoding/hex"
 	"encoding/json"
 	"fmt"
-	"io"
 	"log"
 	"math/big"
 	"time"
@@ -16,21 +14,11 @@ import (
 	cryptorand "crypto/rand"
 
 	"github.com/ethereum/go-ethereum/accounts/keystore"
-	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/ethereum/go-ethereum/crypto/secp256k1"
 	"github.com/pkg/errors"
 	"golang.org/x/crypto/curve25519"
 )
-
-// public properties of key bundle
-type OnChainSigningAddress common.Address
-type OnChainPublicKey ecdsa.PublicKey
-type OffChainPublicKey ed25519.PublicKey
-
-// private properties of key bundle
-type onChainPrivateKey ecdsa.PrivateKey
-type offChainPrivateKey ed25519.PrivateKey
 
 // OCRPrivateKey represents the bundle of keys needed for OCR
 type OCRPrivateKey struct {
@@ -50,9 +38,7 @@ type EncryptedOCRPrivateKey struct {
 	UpdatedAt             time.Time
 }
 
-type ScryptParams struct{ N, P int }
-
-type ocrPrivateKeysRawData struct {
+type ocrPrivateKeyRawData struct {
 	EcdsaX             big.Int
 	EcdsaY             big.Int
 	EcdsaD             big.Int
@@ -60,67 +46,16 @@ type ocrPrivateKeysRawData struct {
 	OffChainEncryption [curve25519.ScalarSize]byte
 }
 
-var defaultScryptParams = ScryptParams{
+type scryptParams struct{ N, P int }
+
+var defaultScryptParams = scryptParams{
 	N: keystore.StandardScryptN, P: keystore.StandardScryptP}
 
 var curve = secp256k1.S256()
 
-func (addr OnChainSigningAddress) Value() (driver.Value, error) {
-	byteArray := [common.AddressLength]byte(addr)
-	return byteArray[:], nil
-}
-
-func (addr *OnChainSigningAddress) Scan(value interface{}) error {
-	switch typed := value.(type) {
-	case []byte:
-		if len(typed) != common.AddressLength {
-			return errors.New("wrong number of bytes to scan into address")
-		}
-		copy(addr[:], typed)
-		return nil
-	default:
-		return errors.Errorf(`unable to convert %v of %T to OnChainSigningAddress`, value, value)
-	}
-}
-
-// Sign returns the signature on msgHash with k
-func (k *onChainPrivateKey) Sign(msg []byte) (signature []byte, err error) {
-	sig, err := crypto.Sign(onChainHash(msg), (*ecdsa.PrivateKey)(k))
-	return sig, err
-}
-
-func onChainHash(msg []byte) []byte {
-	return crypto.Keccak256(msg)
-}
-
-func (k OnChainPublicKey) Address() OnChainSigningAddress {
-	return OnChainSigningAddress(crypto.PubkeyToAddress(ecdsa.PublicKey(k)))
-}
-
-func (k onChainPrivateKey) Address() OnChainSigningAddress {
-	return OnChainSigningAddress(crypto.PubkeyToAddress(k.PublicKey))
-}
-
-// Sign returns the signature on msgHash with k
-func (k *offChainPrivateKey) Sign(msg []byte) ([]byte, error) {
-	if k == nil {
-		return nil, errors.Errorf("attempt to sign with nil key")
-	}
-	return ed25519.Sign(ed25519.PrivateKey(*k), msg), nil
-}
-
-// PublicKey returns the public key which commits to k
-func (k *offChainPrivateKey) PublicKey() OffChainPublicKey {
-	return OffChainPublicKey(ed25519.PrivateKey(*k).Public().(ed25519.PublicKey))
-}
-
 // NewOCRPrivateKey makes a new set of OCR keys from cryptographically secure entropy
 func NewOCRPrivateKey() (*OCRPrivateKey, error) {
-	return newPrivateKey(cryptorand.Reader)
-}
-
-// For internal use only - used to generate new sets of OCR private keys
-func newPrivateKey(reader io.Reader) (*OCRPrivateKey, error) {
+	reader := cryptorand.Reader
 	onChainSk, err := cryptorand.Int(reader, crypto.S256().Params().N)
 	if err != nil {
 		return nil, err
@@ -199,14 +134,6 @@ func (pk *OCRPrivateKey) PublicKeyConfig() [curve25519.PointSize]byte {
 	return rvFixed
 }
 
-// type is added to the beginning of the passwords for
-// OCR keys, so that the keys can't accidentally be mis-used
-// in the wrong place
-func adulteratedPassword(auth string) string {
-	s := "ocrkey" + auth
-	return s
-}
-
 // Encrypt combines the OCRPrivateKey into a single json-serialized
 // bytes array and then encrypts
 func (pk *OCRPrivateKey) Encrypt(auth string) (*EncryptedOCRPrivateKey, error) {
@@ -215,7 +142,9 @@ func (pk *OCRPrivateKey) Encrypt(auth string) (*EncryptedOCRPrivateKey, error) {
 
 // encrypt combines the OCRPrivateKey into a single json-serialized
 // bytes array and then encrypts, using the provided scrypt params
-func (pk *OCRPrivateKey) encrypt(auth string, scryptParams ScryptParams) (*EncryptedOCRPrivateKey, error) {
+// separated into a different function so that scryptParams can be
+// weakened in tests
+func (pk *OCRPrivateKey) encrypt(auth string, scryptParams scryptParams) (*EncryptedOCRPrivateKey, error) {
 	marshalledPrivK, err := json.Marshal(&pk)
 	if err != nil {
 		return nil, err
@@ -263,7 +192,7 @@ func (encKey *EncryptedOCRPrivateKey) Decrypt(auth string) (*OCRPrivateKey, erro
 
 // MarshalJSON marshals the private keys into json
 func (pk *OCRPrivateKey) MarshalJSON() ([]byte, error) {
-	rawKeyData := ocrPrivateKeysRawData{
+	rawKeyData := ocrPrivateKeyRawData{
 		EcdsaX:             *pk.onChainSigning.X,
 		EcdsaY:             *pk.onChainSigning.Y,
 		EcdsaD:             *pk.onChainSigning.D,
@@ -275,7 +204,7 @@ func (pk *OCRPrivateKey) MarshalJSON() ([]byte, error) {
 
 // UnmarshalJSON constructs OCRPrivateKey from raw json
 func (pk *OCRPrivateKey) UnmarshalJSON(b []byte) (err error) {
-	var rawKeyData ocrPrivateKeysRawData
+	var rawKeyData ocrPrivateKeyRawData
 	err = json.Unmarshal(b, &rawKeyData)
 	if err != nil {
 		return err
@@ -308,4 +237,12 @@ func (pk OCRPrivateKey) String() string {
 // GoStringer reduces the risk of accidentally logging the private key
 func (pk OCRPrivateKey) GoStringer() string {
 	return pk.String()
+}
+
+// type is added to the beginning of the passwords for
+// OCR keys, so that the keys can't accidentally be mis-used
+// in the wrong place
+func adulteratedPassword(auth string) string {
+	s := "ocrkey" + auth
+	return s
 }
